requirements.txt
python-dotenv==1.0.0
pyTelegramBotAPI==4.14.0
openai==1.3.0
logging==0.4.9.6 